Make `modify_library_option` backwards compatible
Previously modifying a library option with a new version (>=v3.0.0)
would break old libraries (<v3.0.0)

This was because of the storage overrides introduced in v3.0.0.

This change moves the logic to modify a library option to C++
and it does the protobuf modification without applying a storage
override.

This makes storage overrides work well with both old libraries (as
demonstrated by now passing compatibility tests) and with new libraries
(as confirmed by test_arctic_library_management.py)