Allow setting the on_lock_lost after construction
This will be useful if we want to construct a ReliableStorageLockGuard
but decide later what to do in case a lock is lost.

Also changes ReliableStorageLockGuard to hold a copy of the lock instead
of just a reference. The lock is a cheap object to copy, so it makes
easier to avoid errors where lock is destructed before the guard for it.