Upgrade manylinux to g++-11.5
We were using g++-10 for the manylinux2014 image. This commit
upgrades the CI image to use GCC 11.5.

Discussions with QS lead to deciding that we should upgrade compiler to
gcc 11. But the version availaible in package manager (11.2) had issues
with sparrow. Thus we compile the most recent version of gcc11 (11.5.0)
on top of the manylinux images.