Fix compatibility test cleanup
Previously in several places we were calling `old_ac.clear()` to clear
all arctic/libraries created by an compatibility tests.

This however was producing issues when running pytest with
multiprocessing as we do on the CI:
- Each pytest worker sets up its own instance of the session scoped
  fixtures
- When one process finishes with a fixture it runs the cleanup on an
  arctic instance which is potentially actively used by another pytest
worker. Thus the spurious mongo library not found errors.

To mitigate this we introduce a new context manager `CompatLibrary`. It
is responsible for setting up the `Venv`, `VenvArctic` and `VenvLib` and
also creates the library on `__enter__` and cleans the library on `__exit__`.

So we rework all compat tests to use the new `CompatLibrary`