Don't fail on lmdb storage exceptions while deleting library
For LMDB on windows deleting a directory will fail if a file in it is
still in use. Likely the short lived venv process or an active arctic
instance remains alive for a while longer and prevents the delete.

In this case we just catch the storage exception and leave the deletion
to the temporary direcory used for the root of lmdb.