Skip lmdb compat tests
We have seen spuirous very rare segfaults during cleanup of lmdb compat
tests.

In general lmdb compat tests are not that useful as not much old data is
stored on lmdb. And it is hard to correctly handle lmdb for compat tests
which require muliple processes acessing the same LMDB storage (which
LMDB doesn't fully support)